# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tabdeel - Barter Exchange Platform API for Jordan
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update own profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/listings:
    get:
      operationId: getUserListings
      tags: [users]
      summary: Get user's listings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
  /users/{id}/ratings:
    get:
      operationId: getUserRatings
      tags: [users]
      summary: Get user ratings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRatingsResponse"

  # ─── CATEGORIES ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, phone, city, avatarUrl, rating, ratingCount, isAdmin, isBanned, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        ratingCount:
          type: integer
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, name, city, avatarUrl, rating, ratingCount, listingsCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        ratingCount:
          type: integer
        listingsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        avatarUrl:
          type: string
    UserRatingsResponse:
      type: object
      required: [ratings, avgRating, totalCount]
      properties:
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/Rating"
        avgRating:
          type: number
        totalCount:
          type: integer

    # CATEGORIES
    Category:
      type: object
      required: [id, name, slug, icon]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        listingsCount:
          type: integer

    # LISTINGS
    Listing:
      type: object
      required: [id, title, description, condition, wantsInExchange, city, images, category, owner, isFeatured, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        condition:
          type: string
          enum: [new, used, good]
        wantsInExchange:
          type: string
        city:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          $ref: "#/components/schemas/Category"
        owner:
          $ref: "#/components/schemas/UserProfile"
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Rating:
      type: object
      required: [id, rater, rated, stars, createdAt]
      properties:
        id:
          type: integer
        rater:
          $ref: "#/components/schemas/UserProfile"
        rated:
          $ref: "#/components/schemas/UserProfile"
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...